requirements.txt
flask
mysql-connector-python
python-dotenv
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
reportlab
pytz